- Marten fixed a small error on MSVC in gradient.cpp
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23942 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/aws2/gradient.cpp
@@ -155,7 +155,7 @@ RenderToTexture(JSContext *cx, JSObject *obj, uintN argc, jsval *argv, jsval *rv
 							int dx = mw - x,
 								dy = mh - y;
 								
-							int dist = (int)sqrt((dx*dx) + (dy*dy));
+							int dist = (int)sqrtf((dx*dx) + (dy*dy));
 							
 							if (dist>rl) dist=rl;
